feat: Use mcap compression instead of ros compression
This commit switches from message based compression to the mcap internal compression.
This is done because the ros message compression cannot be replayed without the associated metadata.yaml.
It could not be replayed in external tools like foxglove because they do not load the metadata.yaml.

Using mcap's internal compression compresses just as good and just as fast but has the benefit that it can be opened in all tools.